fix: fixed some bugs

+++ bungee/src/main/java/io/github/milesreimann/clansystem/bungee/command/ClanRoleDeleteSubCommand.java
@@ -3,6 +3,7 @@ package io.github.milesreimann.clansystem.bungee.command;
 import io.github.milesreimann.clansystem.api.entity.ClanMember;
 import io.github.milesreimann.clansystem.api.model.ClanPermissionType;
 import io.github.milesreimann.clansystem.api.service.ClanRoleService;
+import io.github.milesreimann.clansystem.api.service.ClanService;
 import io.github.milesreimann.clansystem.bungee.plugin.ClanSystemPlugin;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 
@@ -15,10 +16,12 @@ import java.util.concurrent.CompletionStage;
  */
 public class ClanRoleDeleteSubCommand extends ClanRoleCommand {
     private final ClanRoleService clanRoleService;
+    private final ClanService clanService;
 
     public ClanRoleDeleteSubCommand(ClanSystemPlugin plugin) {
         super(ClanPermissionType.DELETE_ROLE);
         clanRoleService = plugin.getClanRoleService();
+        clanService = plugin.getClanService();
     }
 
     @Override
@@ -31,15 +34,32 @@ public class ClanRoleDeleteSubCommand extends ClanRoleCommand {
         String name = args[0];
         long clanId = clanMember.getClan();
 
-        return clanRoleService.getRoleByClanIdAndName(clanId, name)
-            .thenCompose(clanRole -> {
-                if (clanRole == null) {
-                    player.sendMessage("rolle gibts nicht");
+        return clanService.getClanById(clanId)
+            .thenCompose(clan -> {
+                if (clan == null) {
                     return CompletableFuture.completedStage(null);
                 }
 
-                return clanRoleService.deleteRole(clanRole)
-                    .thenRun(() -> player.sendMessage("rolle gelöscht"));
+                return clanRoleService.getRoleByClanIdAndName(clanId, name)
+                    .thenCompose(clanRole -> {
+                        if (clanRole == null) {
+                            player.sendMessage("rolle gibts nicht");
+                            return CompletableFuture.completedStage(null);
+                        }
+
+                        if (clan.getOwnerRole().equals(clanRole.getId())) {
+                            player.sendMessage("owner rolle kann nicht gelöscht werden");
+                            return CompletableFuture.completedStage(null);
+                        }
+
+                        if (clan.getDefaultRole().equals(clanRole.getId())) {
+                            player.sendMessage("die standardrolle kann nicht gelöscht werden. setze eine andere rolle als standard");
+                            return CompletableFuture.completedStage(null);
+                        }
+
+                        return clanRoleService.deleteRole(clanRole)
+                            .thenRun(() -> player.sendMessage("rolle gelöscht"));
+                    });
             });
     }
 }
